Add Update To-Do functionality and enhanced readability
Added the functionality to update To-Do tasks by adding an 'Update' button in the listToDos.jsp file. This newly added functionality is supported by the new findById() method in the ToDoService, which enables retrieval of a specific To-Do task by its ID.

Also, changes have been made to improve the readability of the code where the ToDo tasks are being added. Instead of adding all the parameters in one line, they are now split into different lines for better clarity. Furthermore, spacing has been added around '=' for uniformity of code.

The changes are crucial for providing a more interactive and user-friendly experience, as well as for maintaining a cleaner and more readable codebase.