Add date-picker feature and update To-Do functionality
Implemented date-picker functionality in to-do form. A date picker tool has been added to 'todo.jsp' to help users pick a date without having to manually type it. The chosen date format is set as 'yyyy-MM-dd' in application.properties. The webjars bootstrap-datepicker dependency was added to pom.xml to support this functionality.

Also added the feature to update a to-do in 'ToDoService.java' and 'ToDoController.java'. A new method updateToDo(@Valid ToDo todo) has been added in ToDoService.java for updating a todo. This is connected to a new PostMapping in ToDoController.java.

These changes will enhance the user experience by making it easier to select a target date and modify existing to-dos.